+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Feat: OkHttp callback for Customising the Span (#1478)
 * Feat: Add breadcrumb in Spring RestTemplate integration (#1481)
 * Fix: Cloning Stack (#1483)